requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-phonenumber-field==5.2.0
drf-yasg==1.20.0
psycopg2-binary==2.9.1
phonenumbers==8.12.29
requests==2.26.0